+++ packages/xchain-cosmos/src/client.ts
@@ -62,15 +62,21 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
     clientUrls = getDefaultClientUrls(),
     chainIds = getDefaultChainIds(),
     rootDerivationPaths = getDefaultRootDerivationPaths(),
+    customRequestHeaders = {},
   }: XChainClientParams & CosmosClientParams) {
-    super(Chain.Cosmos, { network, rootDerivationPaths, phrase })
+    super(Chain.Cosmos, { network, rootDerivationPaths, phrase, customRequestHeaders })
 
     this.clientUrls = clientUrls
     this.chainIds = chainIds
 
+    if (this.clientUrls[Network.Mainnet].includes('ninerealms.com') && !this.customRequestHeaders['x-client-id']) {
+      this.customRequestHeaders['x-client-id'] = 'xchainjs-client'
+    }
+
     this.sdkClient = new CosmosSDKClient({
       server: this.clientUrls[network],
       chainId: this.chainIds[network],
+      headers: this.customRequestHeaders,
     })
   }